Auto-Correct DSA PROJECT FA19-BCE-086
This report presents and overviews the main working and all the analysis of the Auto-Correct project. This code took a lot of detailed knowledge of file handling. In this project, there is a diversity of knowledge due to use of built-in commands, buffer and file queries, Stacks, Arrays, ArrayList, Inheritance and basic For and While loops. This project serves as an entry level auto-correct or as a suggester. User is allowed to use two files. First file is a file that contains around 5000-7000 words of English grammar, and the other file contains those words which we want to testify for the suggestions/auto correct. Then using some advance level file handling, the code compares both the files. Firstly, a word is taken from test.txt file and that word is being contained in a String data type variable which is then further gone through custom made functions like load words from file, assembles them using a stack and then starts the file handling procedure where all the phenomena take place. For any loop, we find out the runtime of the block inside them and multiply it by the number of times the program will repeat the loop. All loops that grow proportionally to the input size have a linear time complexity O(n). If you loop through only half of the array, that's still O(n). It has a time complexity of O(n) for each word because it must go through all the words that can be referred as ‘n’.